refactor: component my-shop 파일 정리
